Object Managers (layers manager, tetures manager, colliders handler)

+++ Airplane/Airplane/AnimationHandler.cs
@@ -67,7 +67,7 @@ namespace Airplane
             }
         }
 
-        public double Count()
+        public long Count()
         {
             return objectslist_.Count();
         }
